Backend/requirements.txt
fastapi
uvicorn
pypdf
langchain
faiss-cpu
requests
python-multipart
python-dotenv
sentence-transformers
google-generativeai